Fixes to rule conversion into display format (#41)
* Fix parsing of syscall rule with non-specified arch

The new `rule.ToCommandLine` failed when parsing a syscall rule without
an explicit platform.

* Fix test panic when GetStatus returns an error

TestAuditClientSetBacklogWaitTime would panic when fetch of auditd
status fails, for example when run without CAP_AUDIT_READ.

* Fix test build in 32-bit platform

The test was using some constants too big for a 32-bit signed integer.

* Fixes to rule parsing in 32-bit arch

Arch resolution logic was not working properly for 32bit architectures.

Updated the unit tests for 32bit platforms too.

* Tests fixes and exclusions for 32bit platforms

- Fix build error in test
- Exclude golden tests for now (rules not compatible with 32bit)

* Travis: testing round on a 32bit binary

+++ rule/rule_test.go
@@ -362,7 +362,7 @@ func TestAddFilter(t *testing.T) {
 		}
 		assert.EqualValues(t, arg3Field, rule.fields[0])
 		assert.EqualValues(t, equalOperator, rule.fieldFlags[0])
-		assert.EqualValues(t, math.MaxUint32, rule.values[0])
+		assert.EqualValues(t, uint32(math.MaxUint32), rule.values[0])
 	})
 
 	t.Run("arg_min_int32", func(t *testing.T) {
